Restructured app: main.py has app logic; models is a package now
