+++ manage.py
@@ -1,10 +1,9 @@
 import os
 from flask_script import Manager # class for handling a set of commands
 from flask_migrate import Migrate, MigrateCommand
-from app import db, create_app
-from app import models
+from main import db, app
 
-app = create_app(config_name=os.getenv('APP_SETTINGS'))
+#app = create_app(config_name=os.getenv('APP_SETTINGS'))
 migrate = Migrate(app, db)
 manager = Manager(app)
